Reformatting & Restructuring Aggregate Queries
